# contract/openapi.yaml
# GENERATED FILE — DO NOT EDIT.
#
# Written by services/ingestion/cmd/gen-openapi from the Go types in
# libs/wire that the handlers actually decode into, plus the two
# enumerations that live as Go values (policy.ReasonCodes and
# decision.ValidOutcomes).
#
# To change this file, change those types and run:
#
#     make openapi
#
# `make openapi-sync` fails if this file differs from a fresh
# generation, so the specification cannot describe an API the service
# does not serve.

components:
  schemas:
    ClientHints:
      additionalProperties: false
      description: ClientHints are self-reported environment properties.
      properties:
        pointer:
          description: Primary pointing device — the CSS pointer media query as the
            browser reports it
          enum:
          - fine
          - coarse
          - none
          type: string
        reduced_motion:
          description: 'Whether prefers-reduced-motion is set. Load-bearing: it legitimately
            changes movement dynamics.'
          type: boolean
        touch:
          description: Whether the device exposes a touchscreen
          type: boolean
        tz_offset:
          description: Timezone offset in minutes from UTC
          type: integer
        viewport:
          description: Viewport as [width height] in CSS pixels
          items:
            type: integer
          maxItems: 2
          minItems: 2
          type: array
      type: object
    CollectPolicy:
      additionalProperties: false
      description: CollectPolicy is the server-driven collection configuration.
      properties:
        batch_ms:
          description: How often the SDK flushes a batch — milliseconds
          type: integer
        pointer_hz:
          description: Pointer sampling rate in hertz
          type: integer
        types:
          description: Event families the SDK should collect
          items:
            type: string
          type: array
      required:
      - pointer_hz
      - batch_ms
      - types
      type: object
    DecisionEvidence:
      additionalProperties: false
      description: DecisionEvidence is how much was actually observed.
      properties:
        duration_ms:
          description: Wall time those events span — milliseconds
          type: integer
        events:
          description: Interaction events observed in the session
          type: integer
      required:
      - events
      - duration_ms
      type: object
    DecisionReason:
      additionalProperties: false
      description: DecisionReason is one contributing factor.
      properties:
        code:
          description: Stable reason code. Adding a code is non-breaking; changing
            what an existing code means is not (§7).
          enum:
          - POINTER_LINEARITY
          - KEY_INTERVAL_VARIANCE_LOW
          - KEY_DWELL_ABSENT
          - KEY_INTERVALS_IDENTICAL
          - PROGRAMMATIC_SCROLL
          - NO_FOCUS_TRANSITIONS
          - VALUE_INJECTED
          - INSUFFICIENT_EVIDENCE
          type: string
        weight:
          description: Contribution of this factor to the score
          maximum: 1
          minimum: 0
          type: number
      required:
      - code
      - weight
      type: object
    DecisionsRequest:
      additionalProperties: false
      description: DecisionsRequest asks for a judgement.
      properties:
        action:
          description: Action being judged; required. Free-form — for example login
            or checkout.
          type: string
        session_token:
          description: Token returned by POST /v1/sessions
          type: string
        subject_id:
          description: Application's identifier for the actor. Never accepted from
            a browser.
          type: string
      required:
      - action
      type: object
    DecisionsResponse:
      additionalProperties: false
      description: DecisionsResponse is the judgement.
      properties:
        confidence:
          description: How much evidence backs the score. Low confidence gates blocking
            (§3).
          maximum: 1
          minimum: 0
          type: number
        decision:
          description: The decision actually returned. In monitor mode this is always
            allow
          enum:
          - allow
          - challenge
          - block
          type: string
        evaluation_id:
          description: Identifier of this evaluation. Pass it to POST /v1/outcomes
            to label the decision later.
          type: string
        evidence:
          $ref: '#/components/schemas/DecisionEvidence'
        mode:
          description: Operating mode of the service
          enum:
          - monitor
          - enforce
          type: string
        reasons:
          description: Contributing factors. Present so a decision can be explained
            and contested.
          items:
            $ref: '#/components/schemas/DecisionReason'
          type: array
        score:
          description: How bot-like the interaction looks. Rounded to three decimals.
          maximum: 1
          minimum: 0
          type: number
        shadow_decision:
          description: 'In monitor mode: what enforce mode WOULD have returned. Absent
            in enforce mode'
          enum:
          - allow
          - challenge
          - block
          type: string
      required:
      - evaluation_id
      - decision
      - score
      - confidence
      - reasons
      - evidence
      - mode
      type: object
    ErrorResponse:
      additionalProperties: false
      description: ErrorResponse is the single error body shape.
      properties:
        error:
          description: Human-readable failure reason. Not a stable enumeration; do
            not branch on it.
          type: string
      required:
      - error
      type: object
    OutcomesRequest:
      additionalProperties: false
      description: OutcomesRequest labels a past evaluation with what actually happened.
      properties:
        evaluation_id:
          description: evaluation_id returned by POST /v1/decisions; required
          type: string
        observed_at:
          description: RFC 3339 timestamp of the application's observation. Optional;
            rejected with 400 if present and unparseable. The gap between this and
            the server's recorded_at is itself a signal; substituting the server clock
            would silently corrupt it.
          format: date-time
          type: string
        outcome:
          description: 'What the action turned out to be. An unrecognised value is
            rejected rather than stored: a typo''d label silently degrades every future
            calibration.'
          enum:
          - abandoned
          - challenge_failed
          - challenge_passed
          - fraud_confirmed
          - login_failure
          - login_success
          - user_appealed
          type: string
      required:
      - evaluation_id
      - outcome
      type: object
    PageRef:
      additionalProperties: false
      description: PageRef locates the page the session belongs to.
      properties:
        path:
          description: Path of the page — no origin and no query string
          type: string
      type: object
    SessionsRequest:
      additionalProperties: false
      description: SessionsRequest opens a session.
      properties:
        client:
          $ref: '#/components/schemas/ClientHints'
        page:
          $ref: '#/components/schemas/PageRef'
        site_key:
          description: Public tenant key embedded in the page. Identifies the tenant;
            it does not authenticate.
          type: string
      required:
      - site_key
      type: object
    SessionsResponse:
      additionalProperties: false
      description: SessionsResponse carries the session token and the collection policy
        the SDK should apply.
      properties:
        collect:
          $ref: '#/components/schemas/CollectPolicy'
        expires_in:
          description: Seconds until the token stops being accepted
          type: integer
        session_token:
          description: Opaque token correlating telemetry to this session. Not a credential
            and not an identity.
          type: string
      required:
      - session_token
      - expires_in
      - collect
      type: object
    TelemetryBatch:
      additionalProperties: false
      description: TelemetryBatch is one flush from the browser SDK.
      properties:
        events:
          description: The interaction events in this flush. Order is the order observed.
          items:
            $ref: '#/components/schemas/TelemetryEvent'
          type: array
        page:
          $ref: '#/components/schemas/TelemetryPage'
        sent_at_ms:
          description: Milliseconds since session start when the batch was flushed
          type: integer
        seq:
          description: Monotonic batch counter within the session. Used to detect
            loss rather than to order.
          type: integer
        session_token:
          description: 'Session token. May instead travel as an Authorization: Bearer
            header; the header wins when both are present.'
          type: string
      type: object
    TelemetryEvent:
      additionalProperties: false
      description: TelemetryEvent is one interaction event.
      properties:
        action:
          description: 'form: what happened to the field. ''injected'' is the strongest
            single bot signal — a value that appeared with no keystroke and no paste
            behind it; ''paste'' exists so a human pasting is not mistaken for one'
          enum:
          - paste
          - autofill
          - injected
          - submit
          type: string
        class:
          description: 'key: COARSE key class. Never the key itself — no keylogging
            (§2 and §6)'
          enum:
          - alpha
          - digit
          - nav
          - mod
          - whitespace
          - other
          type: string
        dy:
          description: 'scroll: vertical delta in CSS pixels'
          type: integer
        mode:
          description: 'scroll: whether a real wheel gesture produced it or the page
            scrolled itself'
          enum:
          - wheel
          - programmatic
          type: string
        phase:
          description: 'key: keystroke phase. Dwell is down to up on one key; flight
            is up to down between keys'
          enum:
          - down
          - up
          type: string
        pts:
          description: 'pointer: samples as [x y t_offset_ms] triples'
          items:
            items:
              type: integer
            maxItems: 3
            minItems: 3
            type: array
          type: array
        src:
          description: 'pointer: input source that produced the samples'
          type: string
        state:
          description: 'focus or visibility: the state entered. focus and blur for
            focus events; visible and hidden for visibility'
          enum:
          - focus
          - blur
          - visible
          - hidden
          type: string
        t:
          description: Milliseconds since session start
          type: integer
        target:
          description: 'key or form: coarse field role — never the field value'
          type: string
        type:
          description: Event family; selects which of the fields below apply
          enum:
          - pointer
          - key
          - scroll
          - focus
          - visibility
          - form
          type: string
      type: object
    TelemetryPage:
      additionalProperties: false
      description: TelemetryPage is the page state at flush time.
      properties:
        path:
          description: Path of the page — no origin and no query string
          type: string
        viewport:
          description: Viewport as [width height] in CSS pixels
          items:
            type: integer
          maxItems: 2
          minItems: 2
          type: array
      type: object
  securitySchemes:
    secretKey:
      description: The tenant's secret_key. Authenticates the APPLICATION SERVER.
        Compared in constant time, because a byte-wise comparison leaks prefix length
        through response timing on an internet-facing endpoint.
      scheme: bearer
      type: http
    sessionToken:
      description: A session_token from POST /v1/sessions. Correlates telemetry to
        a session; it is NOT a credential and identifies nobody.
      scheme: bearer
      type: http
info:
  description: |-
    Behavioural bot detection from interaction dynamics — how the pointer
    moves, how keys are timed, how a form is actually filled. Not
    fingerprinting: this API makes no claim about which browser is calling.

    GENERATED FILE. Written by services/ingestion/cmd/gen-openapi from the
    Go types the handlers decode into; CI fails if it drifts from them. Edit
    the types, then run `make openapi`.

    Two callers, two trust levels. The BROWSER opens a session and posts
    telemetry, and everything it sends is treated as hostile. The
    APPLICATION SERVER asks for decisions and reports outcomes using
    secret_key; it is the only caller permitted to assert who the subject
    is. A session token correlates telemetry — it does not authenticate
    anyone.

    The service is fail-open. Telemetry is fire-and-forget and loss is
    expected; a decision is a judgement under uncertainty, and score and
    confidence are separate numbers precisely so that "nothing suspicious
    observed" is distinguishable from "this looks human".
  license:
    identifier: Apache-2.0
    name: Apache-2.0
  summary: Does this session behave like a human?
  title: Ghost Trace
  version: 1.0.0
openapi: 3.1.0
paths:
  /healthz:
    get:
      description: Returns 200 while the process is serving. Used by the container
        health check, which execs the binary itself because the runtime image is distroless
        and has no shell.
      operationId: healthz
      responses:
        "200":
          content:
            text/plain:
              example: |
                ok
              schema:
                type: string
          description: Serving
      security: []
      summary: Liveness
      tags:
      - operations
  /v1/decisions:
    post:
      description: |-
        The only endpoint that accepts subject_id and action, which is exactly why neither is ever read from a browser request.

        In monitor mode `decision` is always allow and `shadow_decision` carries what enforce mode would have returned; in enforce mode `shadow_decision` is absent. A client that reads only `decision` therefore measures nothing while the service is in monitor mode.
      operationId: decide
      requestBody:
        content:
          application/json:
            example:
              action: login
              session_token: st_AufHcXG3MEt9x5F3hzVf03ZS
              subject_id: user_8f21
            schema:
              $ref: '#/components/schemas/DecisionsRequest'
        description: Which session, and what it is trying to do.
        required: true
      responses:
        "200":
          content:
            application/json:
              example:
                confidence: 0.448
                decision: allow
                evaluation_id: ev_NORMALIZED
                evidence:
                  duration_ms: 16840
                  events: 120
                mode: monitor
                reasons:
                - code: POINTER_LINEARITY
                  weight: 1
                score: 1
                shadow_decision: challenge
              schema:
                $ref: '#/components/schemas/DecisionsResponse'
          description: A judgement
        "400":
          content:
            application/json:
              example:
                error: action is required
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Malformed JSON body, or action missing
        "401":
          content:
            application/json:
              example:
                error: invalid secret_key
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing or invalid secret_key
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Decision could not be produced
      security:
      - secretKey: []
      summary: Ask for a judgement
      tags:
      - application
  /v1/outcomes:
    post:
      description: |-
        The labels channel every future calibration depends on.

        observed_at is the application's claim about when it saw the outcome; the server separately records its own observation time. The gap between them is itself a signal, so an unparseable observed_at is rejected rather than quietly replaced by the server clock — that substitution would collapse the gap to zero and corrupt the channel invisibly.
      operationId: recordOutcome
      requestBody:
        content:
          application/json:
            example:
              evaluation_id: ev_5Kq2mXbT9vHs
              observed_at: "2026-08-04T09:15:00Z"
              outcome: fraud_confirmed
            schema:
              $ref: '#/components/schemas/OutcomesRequest'
        description: What actually happened to a past evaluation.
        required: true
      responses:
        "202":
          description: Outcome recorded. No body.
        "400":
          content:
            application/json:
              example:
                error: unknown outcome
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Malformed body, missing evaluation_id, unknown outcome, or
            observed_at that is not RFC 3339
        "401":
          content:
            application/json:
              example:
                error: invalid secret_key
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Missing or invalid secret_key
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Outcome could not be recorded
        "503":
          content:
            application/json:
              example:
                error: outcome storage not configured
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 'No durable archive is configured. Refusing is deliberate:
            a label with nowhere to live is worse than a refusal, because the caller
            would believe it had reported one.'
      security:
      - secretKey: []
      summary: Label a past evaluation
      tags:
      - application
  /v1/sessions:
    options:
      description: |-
        Answered for page origins the deployment allows. The allowlist is configuration and never a wildcard; an origin that is not on it receives no Access-Control-Allow-Origin and the browser refuses the request that would have followed.

        Access-Control-Allow-Credentials is never sent: the SDK authenticates with a bearer token it holds in memory and sends no cookies.
      operationId: startSessionPreflight
      responses:
        "204":
          description: Preflight answered. Allow headers are present only for an allowed
            origin.
          headers:
            Access-Control-Allow-Headers:
              description: Content-Type, Authorization.
              schema:
                type: string
            Access-Control-Allow-Origin:
              description: The request's origin, echoed, when it is on the allowlist.
              schema:
                type: string
            Vary:
              description: Always Origin, so a shared cache cannot serve one origin's
                answer to another.
              schema:
                type: string
      security: []
      summary: CORS preflight
      tags:
      - browser
    post:
      description: |-
        Returns a session token and the collection policy the SDK should apply. The collect block is server-driven so collection can be retuned without shipping a new SDK.

        No security scheme: site_key travels in the body. It identifies the tenant and stops cross-tenant noise — it does not authenticate, and it is public in the page source.
      operationId: startSession
      requestBody:
        content:
          application/json:
            example:
              client:
                pointer: fine
                reduced_motion: false
                touch: false
                tz_offset: -180
                viewport:
                - 1440
                - 900
              page:
                path: /login
              site_key: pk_demo
            schema:
              $ref: '#/components/schemas/SessionsRequest'
        description: The tenant's public key plus what the browser reports about itself.
        required: true
      responses:
        "200":
          content:
            application/json:
              example:
                collect:
                  batch_ms: 2000
                  pointer_hz: 20
                  types:
                  - pointer
                expires_in: 1800
                session_token: TOKEN_NORMALIZED
              schema:
                $ref: '#/components/schemas/SessionsResponse'
          description: Session opened
        "400":
          content:
            application/json:
              example:
                error: malformed JSON body
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Malformed JSON body, or a body over 1 MiB
        "401":
          content:
            application/json:
              example:
                error: unknown site_key
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unknown site_key
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Session could not be created
      security: []
      summary: Open a session
      tags:
      - browser
  /v1/telemetry:
    options:
      description: |-
        Answered for page origins the deployment allows. The allowlist is configuration and never a wildcard; an origin that is not on it receives no Access-Control-Allow-Origin and the browser refuses the request that would have followed.

        Access-Control-Allow-Credentials is never sent: the SDK authenticates with a bearer token it holds in memory and sends no cookies.
      operationId: postTelemetryPreflight
      responses:
        "204":
          description: Preflight answered. Allow headers are present only for an allowed
            origin.
          headers:
            Access-Control-Allow-Headers:
              description: Content-Type, Authorization.
              schema:
                type: string
            Access-Control-Allow-Origin:
              description: The request's origin, echoed, when it is on the allowlist.
              schema:
                type: string
            Vary:
              description: Always Origin, so a shared cache cannot serve one origin's
                answer to another.
              schema:
                type: string
      security: []
      summary: CORS preflight
      tags:
      - browser
    post:
      description: |-
        Fire-and-forget. Always 202 on an accepted request, INCLUDING when the session token is unknown or expired: telemetry loss is expected (§5), and an error there would only teach a stale SDK to retry in a loop. Ingestion is idempotent on (session, seq).

        The token may travel as an Authorization: Bearer header or in the body; the header wins.
      operationId: ingestTelemetry
      requestBody:
        content:
          application/json:
            example:
              events:
              - pts:
                - - 412
                  - 300
                  - 0
                - - 418
                  - 305
                  - 16
                - - 427
                  - 311
                  - 17
                src: mouse
                t: 1200
                type: pointer
              - class: alpha
                phase: down
                t: 1500
                target: f_1
                type: key
              - class: alpha
                phase: up
                t: 1530
                target: f_1
                type: key
              page:
                path: /login
                viewport:
                - 1440
                - 900
              sent_at_ms: 4200
              seq: 1
              session_token: st_AufHcXG3MEt9x5F3hzVf03ZS
            schema:
              $ref: '#/components/schemas/TelemetryBatch'
        description: One flush of interaction events. The pts triples are [x, y, milliseconds
          since the previous sample].
        required: true
      responses:
        "202":
          description: Accepted. No body — including for an unknown session.
        "400":
          content:
            application/json:
              example:
                error: malformed JSON body
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Malformed JSON body, or a body over 1 MiB
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Telemetry could not be ingested
      security:
      - sessionToken: []
      - {}
      summary: Post a telemetry batch
      tags:
      - browser
servers:
- description: Local slice (make run, or the compose core profile)
  url: http://127.0.0.1:8080
tags:
- description: Called by the SDK in the page. Hostile input. Answers cross-origin
    requests from an allowlist of page origins; the application endpoints deliberately
    do not.
  name: browser
- description: Called by the application server with secret_key.
  name: application
- description: Liveness.
  name: operations